requirements.txt
appdirs==1.4.0
certifi==2017.4.17
chardet==3.0.4
click==6.7
enum34==1.1.6
Flask==0.12.2
greenlet==0.4.12
idna==2.5
itsdangerous==0.24
Jinja2==2.9.6
kaitaistruct==0.7
MarkupSafe==1.0
packaging==16.8
pyparsing==2.1.10
python-pushover==0.3
requests==2.18.1
six==1.10.0
urllib3==1.21.1
Werkzeug==0.12.2
